move sklearn span analyzer code to a module
